style/routing: have home page and Files page share a layout with sidebar
