+++ src/App.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router';
 import { CookiesProvider } from 'react-cookie';
 import { MainLayout } from './layouts/MainLayout';
+import { FilesLayout } from './layouts/FilesLayout';
 import Home from './components/Home';
 import Files from './components/Files';
 import Help from './components/Help';
@@ -48,12 +49,14 @@ const AppComponent = (): JSX.Element => {
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/*" element={<MainLayout />}>
-          <Route path="files" element={<Files />} />
           <Route path="jobs" element={<Jobs />} />
           <Route path="help" element={<Help />} />
           <Route path="profile" element={<Profile />} />
           <Route path="preferences" element={<Preferences />} />
-          <Route path="*" element={<Home />} />
+          <Route element={<FilesLayout/>}>
+            <Route path="files" element={<Files />} />
+            <Route index path="*" element={<Home />} />
+          </Route>
         </Route>
       </Routes>
     </BrowserRouter>
