+++ src/components/Files.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import FileSidebar from '../components/ui/FileSidebar';
 import FileList from '../components/ui/FileList';
 import useFileBrowser from '../hooks/useFileBrowser';
 
@@ -9,18 +8,9 @@ export default function Files() {
     fileSharePaths,
     currentPath,
     checked,
-    openZones,
     getFiles,
-    getFileSharePaths,
     handleCheckboxToggle,
-    toggleZone
   } = useFileBrowser();
-
-  React.useEffect(() => {
-    if (Object.keys(fileSharePaths).length === 0) {
-      getFileSharePaths();
-    }
-  }, [fileSharePaths, getFileSharePaths]);
 
   React.useEffect(() => {
     if (files.length === 0 && Object.keys(fileSharePaths).length > 0) {
@@ -28,23 +18,10 @@ export default function Files() {
     }
   }, [files, fileSharePaths, getFiles]);
 
-  // Handler for when a path is clicked in the sidebar
-  const handlePathClick = (path: string) => {
-    getFiles('');
-  };
   console.log('files in Files.tsx', files);
   console.log('fileSharePaths in Files.tsx', fileSharePaths);
 
   return (
-    <div className="flex h-full w-full">
-      <div className="w-64 h-full border-r border-gray-200 overflow-y-auto">
-        <FileSidebar
-          fileSharePaths={fileSharePaths}
-          openZones={openZones}
-          toggleZone={toggleZone}
-          onPathClick={handlePathClick}
-        />
-      </div>
       <div className="flex-1 h-full overflow-auto">
         <FileList
           files={files}
@@ -54,6 +31,5 @@ export default function Files() {
           getFiles={getFiles}
         />
       </div>
-    </div>
   );
 }
